+++ R/is_windows.R
@@ -1,6 +1,5 @@
 #' Is the System Running a Windows Machine?
 #'
 #' @return \code{\link{TRUE}} if so, \code{\link{FALSE}} otherwise.
-#' @keywords internal
 #' @export
 is_windows <- function() return(.Platform[["OS.type"]] == "windows")