+++ inst/runit_tests/runit-load_internal_functions.R
@@ -1,8 +1,8 @@
 if (interactive()) pkgload::load_all(".")
 test_load_internal_functions <- function() {
     load_internal_functions("fritools")
-    result <- ls()
-    expectation <- "throw"
+    result <- environmentName(environment(throw))
+    expectation <- "fritools"
     RUnit::checkIdentical(result, expectation)
 }
 if (interactive()) {