+++ NEWS.md
@@ -1,6 +1,18 @@
+# fritools 1.1.0
+
+* Fixed buggy regular expression in `is_running_on_gitlab_com()`
+* Added `get_options`, `set_options`, `is_force`; 
+  `call_conditionally`, `call_safe`;
+  `is_installed`, `is_r_package_installed`;
+  `is_false`, `is_null_or_true`;
+  `search_files`, `find_files`, `summary.filesearch` and
+  `strip_off_attributes`
+  from package `packager`.
+
+
 # fritools 1.0.0
 
-* Get the compilation of utilities from
+* Got the compilation of utilities from
    - rasciidoc/R/utils.R: *
    - packager/R/is\_version\_sufficient.R: *
    - rasciidoc/R/is\_version\_sufficient.R: *